+++ lbh15/properties/bismuth_thermochemical_properties/diffusivity_in_bismuth.py
@@ -4,7 +4,6 @@ from typing import List
 import numpy as np
 from scipy.constants import atm
 from scipy.constants import R
-from lbh15.properties.interface import PropertyInterface
 from lbh15.properties.common_interface import OxygenDiffusivityInterface
 from ..._decorators import range_warning
